% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_create_fimo_predictors.R
\name{addToScoreList}
\alias{addToScoreList}
\title{Adds a score to the score list}
\arguments{
\item{trainingD:}{a table of eQTL causal SNPs}

\item{fimoS:}{a score for FIMO}

\item{scoreName:}{string, name indicating analysis type and type of score
Ex. "fb_minPval"}

\item{scoreList:}{a vector containing all scores for FactorBook, ENCODE,
Jaspar, and Deepbhind}
}
\description{
\code{addToScoreList} adds a FactorBook or Jaspar score to the score list.
  Specifically, minPval, maxScore, numOcc, or sumScore.
}

