% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_construct_overlapping.encodepeaks.narrow1_SS.R
\name{construct_overlapping_encodepeaks}
\alias{construct_overlapping_encodepeaks}
\title{Calculates ENCODE chromatin scores}
\arguments{
\item{inputDTLocation:}{file path of eQTLs to be analyzed (.Rda file)}

\item{outputDir:}{string, directory to save output to}
}
\value{
NULL, but saves results to disk as FeaturesEncodePeaksNarrow1_v3.Rda
}
\description{
\code{construct_overlapping_encodepeaks} takes the metadata peaks files
  generated by extract_encodepeaks_SS.R and ____
}
\details{
Reads in the encodePeaks_v3.Rda file generated by extract_encodepeaks_SS.v3.R
}

