% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_construct_summary.8merscores_ref_SS.R
\name{constructSummary8merScores}
\alias{constructSummary8merScores}
\title{Summarizes results of 8mer scores}
\arguments{
\item{inputDTFile:}{string, location of eQTL data table}

\item{outputDir:}{string, file location to save results}
}
\value{
Null. Saves results to disk in outputDir
}
\description{
\code{constructSummary8merScores} takes the 8mer files from extract_8merscores.R
 and creates a summary
}

