% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_create_fimo_predictors.R
\name{create_fimo_predictors}
\alias{create_fimo_predictors}
\title{Run all functions in here}
\arguments{
\item{trainingDLocation:}{string, location of the training D}

\item{dataDir:}{}

\item{fimoDir:}{}
}
\description{
Run all functions in here
}

