% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_construct_summary.8merscores_ref_SS.R
\name{createEnrscoresList}
\alias{createEnrscoresList}
\title{Imports human and mouse data}
\description{
\code{createEnrscoresList} imports the human and mouse .Rda files created
  in extract_8merscores and creates and labels a list of scores.
}
\details{
Helper function for constructSummary8merScores
}

