% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_construct_summary.8merscores_ref_SS.R
\name{createSubseq21List}
\alias{createSubseq21List}
\title{title}
\description{
\code{createSubseq21List} reads in a DT of eQTL data and produces _______
}
\details{
Helper function for constructSummary8merScores
}

