% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_DFToGranges.R
\name{dt2granges}
\alias{dt2granges}
\title{Converts data.table to Granges}
\arguments{
\item{dt:}{a data.table object}
}
\value{
Granges object containing same information
}
\description{
\code{dt2granges} converts a data.table object to Granges
}
\details{
modified from a function on http://davetang.org/muse/2015/02/04/bed-granges/
}

