% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_DFToGranges.R
\name{dt2uniquegranges}
\alias{dt2uniquegranges}
\title{Converts data.table to UniqueGranges}
\arguments{
\item{dt:}{a data.table object}
}
\value{
Granges object containing unique peaks
}
\description{
\code{dt2granges} converts a data.table object to unique Granges
}

