% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_extract_8merscores_ref_SS.R
\name{extract8mers}
\alias{extract8mers}
\title{Extract all 8mers for given eQTL sequences}
\usage{
extract8mers(uniprobeDir, inputDTFile, outputDir, organismName)
}
\arguments{
\item{datadir:}{string, directory where Uniprobe scores have been saved}

\item{sampleDTLocation:}{string, file path of input data (in .Rdata format)}

\item{outputDir:}{string, directory to save output to}

\item{organismName:}{string, all lowercase, "human" or "mouse}
}
\value{
Returns null, saves output in outputDir/escores8mer_ref_human_v2.Rda
}
\description{
\code{extract8mers} Computes enrichment scores for all 8mers on the 150bp sequences 
  given, output to drive
}
\details{
Calls helper functions from 8merScores.R
}

