% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{extract_encodepeaks}
\alias{extract_encodepeaks}
\title{Extracts peaks from ENCODE data}
\arguments{
\item{chipLocation:}{string, file location of ENCODE ChIP metadata.tsv}

\item{dnaseLocation:}{string, file location of ENCODE DNase metadata.tsv}

\item{dnamethLocation:}{string, file location of ENCODE DNA methylation 
metadata.tsv}

\item{peakdir:}{string, path of directory containing peak files in .bed format}

\item{outdir:}{string, path of directory to save output}
}
\value{
NULL, but saves to disk an .Rda file of
  meta_peaks: a data.table of all the peak metadata
  bed.list.raw: a list of the raw bed file names 
  bed.list.narrow1: a list of all the narrowpeak files
  uniquepeaks.raw: a list of the raw unique peaks
  uniquepeaks.narrow1: a list of the unique narrowpek files
This is what is actually called by run_Encode.R
}
\description{
\code{extract_encode peaks} takes ENCODE data and outputs a list of peaks
  with metadata, as well as lists of bed files and unique peaks.
}

