% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_create_fimo_predictors.R
\name{factorBookAnalysis}
\alias{factorBookAnalysis}
\title{Exploratory analysis with FactorBook PWM scores}
\arguments{
\item{scoreList:}{}
}
\value{

}
\description{
\code{factorBookAnalysis}
}

