% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{getAllLists}
\alias{getAllLists}
\title{Gets normal, unique, and narrowpeak bed files}
\arguments{
\item{peakname:}{}
}
\value{
a data.table of ___
helper fn called by getBedList
}
\description{
\code{getAllLists} separates the metadata of peaks into narrow peaks, broad
  peaks, and methylation data, then returns a variety of lists of bed files,
  including the peaks unique for genomic ranges.
}

