% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{getBedDF}
\alias{getBedDF}
\title{Read in a bed file}
\arguments{
\item{peakname:}{string, filename}
}
\value{
a data.table of ___
helper fn called by getBedList
}
\description{
\code{getBedDF} reads in a bed file and saves it as a data.table
}

