% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{getBedList}
\alias{getBedList}
\title{Turns bed files into data.tables}
\arguments{
\item{bedNames:}{a list(?) of .bed file names}
}
\value{
a list of data.tables
helper fn called by getAllLists
}
\description{
\code{getBedList} gets a list of data.tables containing bed file info
}

