% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{getBedNames}
\alias{getBedNames}
\title{Get name of Bed files for all peaks}
\arguments{
\item{file_type:}{string of file name}

\item{meta_peaks:}{data.table of the ENCODE metadata}
}
\value{
a data.table of strings of all the bed file names
}
\description{
\code{getBedNames} gets a data.ta______
}

