% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_8merScores.R
\name{getCols3to5}
\alias{getCols3to5}
\title{Returns 3rd to 5th columns}
\arguments{
\item{DT}{data.table}
}
\value{
a data.table of up to 3 columns
}
\description{
\code{getCols3to5} returns the 3rd - 5th columns of a data.table.
  If the table has less than 5 columns, it returns 3rd - max.
  If the table has less than 3 columns, it returns a null table.
}

