% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_create_fimo_predictors.R
\name{getCorrelation}
\alias{getCorrelation}
\title{\code{getCorrelation} performs a simple correlation between a fimo score and
   the ????? from the eQTL data given.}
\arguments{
\item{fimoS.ordered, }{???}

\item{trainingD:}{a table of eQTL data}
}
\description{
\code{getCorrelation} performs a simple correlation between a fimo score and
   the ????? from the eQTL data given.
}

