% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_getDefaultFileLocations.R
\name{getDefaultFileLocations}
\alias{getDefaultFileLocations}
\title{Get file locations of Data}
\arguments{
\item{uniprobeDirectory:}{string, location of /Uniprobe scores
Should have /Human and /Mouse subdirectories}

\item{chipLocation:}{string, location of ENCODE ChIP-seq metadata}

\item{dnaseLocation:}{string, location of ENCODE DNase-seq metadata}

\item{dnamethLocation:}{string, location of ENCODE DNA methylation metadata}

\item{peakDir:}{string, directory of ENCODE peaks}

\item{inputDTFile:}{string, location of eQTL data to analyze
Data set from CAGI 2015 challenge is /ProcessedData/SampleDT.Rdata}

\item{outputDir:}{string, directory to save output in}
}
\value{
A list
}
\description{
\code{getDefaultFileLocations} sets and loads file paths of important data.
}
\details{
These locations are only stored here in the package. This is the only place
  that default locations need to be changed. Default locations set for 
  Keles' lab's server.
The ENCODE data are peak/bed files obtained from encodeproject.org
}

