% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_run_fimo_and_summarize.R
\name{getFactorName}
\alias{getFactorName}
\title{Creates directory name for factor}
\arguments{
\item{refOrAlt:}{string of "ref" or "alt"}
}
\description{
\code{getFactorName} creates a directory name for a directory within
  fimo to save to. 
 
Relies on previously defined fimoThreshold
Ex: "Training_ref_thres0.001"
}

