% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_run_fimo_and_summarize.R
\name{getFastaName}
\alias{getFastaName}
\title{Creates filename for FASTA file}
\arguments{
\item{refOrAlt:}{string of "ref" or "alt"}
}
\description{
\code{getFactorName} creates filename
 
Ex: "ref_SS.v1.fasta"
}

