% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{getMetaPeaks}
\alias{getMetaPeaks}
\title{Combines all peak files from ENCODE}
\arguments{
\item{chipLocation:}{string, file location of ENCODE ChIP metadata.tsv}

\item{dnaseLocation:}{string, file location of ENCODE DNase metadata.tsv}

\item{dnamethLocation:}{string, file location of ENCODE DNA methylation 
metadata.tsv}
}
\value{
data.table containing all info from ChIP, DNase, and DNA methylation
}
\description{
\code{readChip} reads in ChIP seq, DNase seq, DNAmethylation and combines
  the peaks into one file
}

