% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_8merScores.R
\name{getRangeCols}
\alias{getRangeCols}
\title{Get range of a data.table's columns}
\arguments{
\item{DT}{data.table}

\item{RangeCols3to5, }{a 3 column data.table}
}
\value{
a data.table with 2 rows and the same number of columns

something
}
\description{
\code{getRangeCols} finds the minimum and maximum of each column in DT, then
  returns a data table with the minimum in the first row and the maximum in
  the second row.

\code{getEscoresInd} does _____
}

