% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_construct_summary.8merscores_ref_SS.R
\name{GetSubseqCoordAbs}
\alias{GetSubseqCoordAbs}
\title{Get absolute coordinates for the subsequences}
\arguments{
\item{subseq.len:}{int,}

\item{center.len:}{}

\item{center.pos:}{variable of positions from inputDTFile}
}
\description{
\code{GetSubseqCoordAbs} takes something and makes something
}
\details{
#' Helper function for createSubseq21List()
}

