% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_create_fimo_predictors.R
\name{loadFimoData}
\alias{loadFimoData}
\title{Load in FIMO summary data}
\value{
Null. Loads 4 .Rdata files into the environment
}
\description{
\code{loadFimoData} loads in the .Rdata files created by 
  fimo_run_fimo_and_summarize.R
}

