% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_create_fimo_predictors.R
\name{makeScoreList}
\alias{makeScoreList}
\title{Creates a score list}
\value{
scoreList: a list which will contain the FactorBook, Jaspar, Encode,
  and Deepbind scores.
}
\description{
\code{makeScoreList} creates and gives column names to a list which will
  contain the minPval, maxScore, numOcc, and sumScore for Factorbook, 
  Jaspar, and Encode scores for the given eQTLs, as well as the Deepbind
  score.
}

