% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{readChip}
\alias{readChip}
\title{read in ChIP data from ENCODE}
\arguments{
\item{chipLocation:}{string, file location of ENCODE ChIP metadata.tsv}
}
\value{
data.table containing the ChIP info
}
\description{
\code{readChip} reads in broad and narrow peak files from ENCODE ChiP metadata
}

