% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{readDnameth}
\alias{readDnameth}
\title{read in DNA methylation data from ENCODE}
\arguments{
\item{chipLocation:}{string, file location of ENCODE DNA methylation 
metadata.tsv}
}
\value{
data.table containing the DNA methylation info
}
\description{
\code{readChip} reads in files from ENCODE DNA 
  methylation metadata
}

