% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{readDnase}
\alias{readDnase}
\title{read in Dnase data from ENCODE}
\arguments{
\item{chipLocation:}{string, file location of ENCODE Dnase metadata.tsv}
}
\value{
data.table containing the Dnase info
}
\description{
\code{readChip} reads in narrow peak files from ENCODE Dnase metadata
}

