% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_8merScores.R
\name{removeHeader}
\alias{removeHeader}
\title{Removes header from a data.table}
\description{
\code{removeHeader} takes the proper number of rows for dt (numRowsNoHeader)
  and if dt has one extra row, it removes the first row (header) from dt.
  If dt does not have numRowsNoHeader rows, or numRowsNoHeader + 1 rows,
  this function returns an error.
}

