% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_run_Encode.R
\name{run_encode}
\alias{run_encode}
\title{ENCODE data}
\value{
Null. Data saved to disk.
}
\description{
\code{run_encode} is the main function for interpreting ChIP, DNase, and DNA methylation data
}
\details{
Relies on output from  uniprobe_run_Uniprobe_scores.R.
Depends on encode_extract_encodePeaks.R, encode_DFToGranges.R, and 
encode_construct_overlapping.encodepeaks.narrow1_SS.R.
Default locations set for Keles' lab and are read in from
  base_getDefaultFileLocations.R.
}

