% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_run_fimo_and_summarize.R
\name{run_fimo_and_summarize}
\alias{run_fimo_and_summarize}
\title{Runs FactorBook, ENCODE and Jaspar on FIMO}
\arguments{
\item{refOrAlt:}{string, "ref" or "alt"}

\item{memeFimoDir:}{string,}

\item{pwmFiles:}{list of strings, locations of PWMs for ENCODE, Factorbook,
and Jaspar'}
}
\value{
Null. Saves to disk.
}
\description{
Runs FactorBook, ENCODE and Jaspar on FIMO
}

