% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprobe_run_Uniprobe.R
\name{run_Uniprobe_scores}
\alias{run_Uniprobe_scores}
\title{Computing Uniprobe 8mer scores}
\usage{
run_Uniprobe_scores()
}
\value{
Null. Data saved to disk.
}
\description{
\code{run_Uniprobe_scores} is the main function for computing Uniprobe 8mer
  scores
}
\details{
Depends on uniprobe_8merScores.R, uniprobe_extract_8merscores_SS.R, and 
  uniprobe_construct_summary.8merscores_ref_SS.R
Default locations set for Keles' lab and are read in from
  base_getDefaultFileLocations.R.
}

