% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_run_fimo_and_summarize.R
\name{runFimo}
\alias{runFimo}
\title{Run Fimo}
\arguments{
\item{memeFimoDir:}{string, path of fimo directory for MEME}

\item{outDir:}{string, location to save to}

\item{pwmFiles:}{list of strings, locations of PWM for ENCODE, Factorbook, and Jaspar}

\item{factorName:}{string, created by getFactorName}

\item{fastaFile:}{string, created by getFastaName}
}
\description{
Sends a system command to run FIMO
}

