% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_extract_encodepeaks.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalizes words in a string}
\arguments{
\item{x, }{a string with words separated by spaces}
}
\description{
\code{simpleCap} capitalizes the first letter of every word in a string
}
\details{
Helper function for readChip, readDnase, and readDnameth
}

