% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo_run_fimo_and_summarize.R
\name{summarizeFactorBook}
\alias{summarizeFactorBook}
\title{Summarize FIMO with FactorBook}
\arguments{
\item{factorBookPWM:}{a ____ of ____}
}
\value{
Null. Saves to disk.
}
\description{
Summarize FIMO with FactorBook
}

